test_mysql_flask/app/requirements.txt
Flask
numpy
pymysql
tensorflow
python-dotenv
mysql-connector-python
